test: user roles and admin permissions to delete boards

+++ backend/src/modules/boards/services/delete.board.service.ts
@@ -8,6 +8,7 @@ import isEmpty from 'libs/utils/isEmpty';
 import { GetTeamServiceInterface } from 'modules/teams/interfaces/services/get.team.service.interface';
 import * as Teams from 'modules/teams/interfaces/types';
 import { TeamUserDocument } from 'modules/teams/schemas/team.user.schema';
+import { UserDocument } from 'modules/users/schemas/user.schema';
 
 import { DeleteBoardService } from '../interfaces/services/delete.board.service.interface';
 import Board, { BoardDocument } from '../schemas/board.schema';
@@ -34,6 +35,26 @@ export default class DeleteBoardServiceImpl implements DeleteBoardService {
 		return teamUser;
 	}
 
+	private async isUserSAdmin(
+		userId: string,
+		quantidadeUsers: LeanDocument<TeamUserDocument>[]
+	): Promise<boolean> {
+		const myUser = quantidadeUsers.find(
+			(user) => String((user.user as UserDocument)?._id) === String(userId)
+		);
+		const isUserSAdmin = (myUser?.user as UserDocument).isSAdmin;
+		return isUserSAdmin;
+	}
+
+	private async getUsersOfTeam(teamId: string): Promise<LeanDocument<TeamUserDocument>[]> {
+		const users = await this.getTeamService.getUsersOfTeam(teamId);
+		if (!users) {
+			throw new NotFoundException('User not found list of users!');
+		}
+
+		return users;
+	}
+
 	async deleteSubBoards(dividedBoards: Board[] | ObjectId[], boardSession: ClientSession) {
 		const { deletedCount } = await this.boardModel
 			.deleteMany({ _id: { $in: dividedBoards } }, { session: boardSession })
@@ -54,8 +75,6 @@ export default class DeleteBoardServiceImpl implements DeleteBoardService {
 	}
 
 	async deleteBoard(boardId: string, userId: string, boardSession: ClientSession) {
-		// console.log('deleteBoard', boardId, userId);
-
 		const result = await this.boardModel.findOneAndRemove(
 			{
 				_id: boardId
@@ -69,32 +88,32 @@ export default class DeleteBoardServiceImpl implements DeleteBoardService {
 
 	async delete(boardId: string, userId: string) {
 		const boardSession = await this.boardModel.db.startSession();
-		// console.log('boardModel', this.boardModel);
-		// const boardTest = await this.boardModel.findOne({
-		// 	_id: boardId
-		// });
-		// console.log('boardTest', boardTest);
 		const board = await this.boardModel.findById(boardId).exec();
 		if (!board) {
 			throw new NotFoundException('Board not found!');
 		}
 		const { team, createdBy } = board;
 		const teamUser = await this.getTeamUser(userId, String(team));
-		console.log('teamuser', teamUser.role);
+		const users = await this.getUsersOfTeam(String(team));
+
+		const userIsSAdmin = await this.isUserSAdmin(userId, users);
+		console.log('userIsAdmin', userIsSAdmin);
+
 		const isAdminOrStakeholder = [TeamRoles.STAKEHOLDER, TeamRoles.ADMIN].includes(
 			teamUser.role as TeamRoles
 		);
-		console.log(isAdminOrStakeholder);
 
 		// Validate if the logged user are the owner
 		const isOwner = String(userId) === String(createdBy);
 
-		console.log('isOwner', isOwner);
+		if (!(isOwner || isAdminOrStakeholder || userIsSAdmin)) {
+			console.log('  NAO PODES ALTERAR');
+			throw Error(DELETE_FAILED);
+		}
 
 		const boardUserSession = await this.boardUserModel.db.startSession();
 		boardSession.startTransaction();
 		boardUserSession.startTransaction();
-
 		try {
 			const { _id, dividedBoards } = await this.deleteBoard(boardId, userId, boardSession);
